app/prototype/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
ddgs==9.13.1
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
lxml==6.0.4
numpy==2.0.2
ollama==0.6.1
openai==2.31.0
pgvector==0.4.2
primp==1.2.2
psycopg2-binary==2.9.11
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.1
regex==2026.1.15
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.49.3
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
